refactor(tarko-agent-ui): merge duplicate replay hooks and eliminate redundancy
